test: seed transaction regressions through canonical semantic lowering
